pyproject.toml
[project]
name = "structured-tutorials"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "sphinx>=7.3.7",
]

[project.scripts]
structured-tutorial = "structured_tutorials.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
source = [
    "structured_tutorials",
]
branch = true

[tool.doc8]
max-line-length = 110
ignore-path = ["docs/_build/"]
ignore = ["D000"]

[tool.mypy]
strict = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
addopts = [
    "--cov",
    "--cov-report=html",
    "--cov-fail-under=100",
    "--random-order",
]

[tool.ruff]
extend-exclude = ["migrations", "stubs"]
line-length = 110

[tool.ruff.lint]
extend-select = [
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "G",  # flake8-logging-format
    "PT",  # flake8-pytest-style
    "UP",  # pyupgrade
    "RUF",  # ruff specific checks
]

ignore = [
    "D102",  # Missing docstring in public method - includes inherited methods
    "D104",  # Missing docstring in public package - includes even empty __init__.py files
    "D105",  # Missing docstring in magic method - includes __str__() etc
    "D106",  # Missing docstring in public nested class - includes Meta/Media classes
    "D107",  # Missing docstring for __init__()
    "D203",  # incorrect-blank-line-before-class
    "D213",  # multi-line-summary-second-line`
    "D301",  # Use `r"""` if any backslashes in a docstring - incompatible with doctests!
    "D401",  # First line of docstring should be in imperative mood
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.uv]
conflicts = [
    [
        { group = "Sphinx7.3" },
        { group = "Sphinx7.4" },
        { group = "Sphinx8.1" },
        { group = "Sphinx8.2" },
    ],
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "doc8>=2.0.0",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-random-order>=1.1.1",
    "pytest-subprocess>=1.5.3",
    "ruff>=0.11.10",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxcontrib-spelling>=8.0.1",
    "types-docutils>=0.21.0.20250604",
    "types-pyyaml>=6.0.12.20250516",
]

"Sphinx7.3" = [ "Sphinx~=7.3.0" ]
"Sphinx7.4" = [ "Sphinx~=7.4.0" ]
"Sphinx8.1" = [ "Sphinx~=8.1.0" ]
"Sphinx8.2" = [ "Sphinx~=8.2.0" ]
